add house-listing search by center and radius

+++ src/main/java/it/unina/dietiestates25/listing/infrastructure/adapter/in/RestListingController.java
@@ -140,7 +140,7 @@ public class RestListingController {
             @RequestParam(required = false) String city,
             @RequestParam(required = false) Double centerLongitude,
             @RequestParam(required = false) Double centerLatitude,
-            @RequestParam(required = false) Double radius,
+            @RequestParam(required = false) Integer radius,
             @RequestParam(required = false) Double priceMin,
             @RequestParam(required = false) Double priceMax,
             @RequestParam(required = false) Double squareMetersMin,
